moved all widgets into their own files
