+++ lib/bureaucrat/widgets/hidden_input.rb
@@ -1,0 +1,15 @@
+require 'bureaucrat/widgets/input'
+
+module Bureaucrat
+  module Widgets
+    class HiddenInput < Input
+      def input_type
+        'hidden'
+      end
+
+      def hidden?
+        true
+      end
+    end
+  end
+end
