+++ lib/bureaucrat/widgets/text_input.rb
@@ -1,0 +1,11 @@
+require 'bureaucrat/widgets/input'
+
+module Bureaucrat
+  module Widgets
+    class TextInput < Input
+      def input_type
+        'text'
+      end
+    end
+  end
+end
